Made it so that it gets more than 100 tracks in a playlist.
Do note that it may take a while to load if the user does have a lot of playlists with a lot of tracks. I have tested around 55 playlists where some playlists may have had around 300 tracks and it took about 10-15 seconds to load. Note that some revisions for error checking may need to be done.