Updated app.py, spotifyData.py, databaseops.py and created spotifyData.js.
app.py has been updated so that it can send data to the webpage to display playlists in a dropdown menu and also to get songs from the specified playlist and anaylze a certian amount of songs from user's recently played spotify songs. The updates done to spotifyData.py are the following:
get_songs_from_playlist: The return value is now a list of dicts which contains song name and artist or list of artists.
spotifyData.js:
playlist function basically updates the widget if the user changes the playlist in the dropdown list and it also sends the id of the playlist to flask, when the analyze button is clicked.
recent_plays: Same as the playlist except it does not send a playlist id instead it use the value inside of the text to know how many recent songs it should get, if no input is given the default number of recent songs is 10.
Note that some testing is still needed.