Updated app.py, spotifyData.js, spotify.css and spotify.css.
In app.py I simply copied parse_songdata and named it as parse_songdata2  to be able to continue on working with data that gets analyzed and displaying the coloring. Now after the user clicks the analyze playlist button a drop down menu containing all the songs that get anaylzed will be genreated and the user can choose which song highlighting they would like to see, these changes are done in spotifyData.js and spotify.html, where spotifyData.js will get data that was returned from flask and create a drop down menu if there isn't one already and set up the options needed. The pSongs function simply updates what is in the output box to match the selected song. Note that I have yet to implement a dropdown menu for the recent plays which should be fairly simple.